requirements.txt
about-time==4.2.1
alive-progress==3.2.0
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
autograd==1.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cma==3.2.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.9
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
gmsh==4.13.1
grapheme==0.6.0
h5py==3.14.0
htmlmin==0.1.12
idkDOE @ git+https://kodea.danobatgroup.com/dip/precision/ideko/simulation/idkdoe.git@46557db57296223536b32215b8e7ae5210e43c60
idkfem @ git+https://kodea.danobatgroup.com/dip/precision/ideko/simulation/idkFEM.git@5ed01477ef7961e0b253c98e7abfe7b67b9e2b87
idkOPT @ git+https://kodea.danobatgroup.com/dip/precision/ideko/simulation/idkopt.git@8459269b543ccc4a3892c6ad8c9958174c1ae4ec
idkrom @ git+https://kodea.danobatgroup.com/dip/precision/ideko/simulation/idkROM.git@aa5806fd0c99b68402d4633b164ce7c7e0358801
idkSIM @ file:///D:/portable_idkSIM/idksimulation
idna==3.10
ImageHash==4.3.1
intel-cmplr-lib-ur==2025.0.5
intel-openmp==2025.0.5
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ismember==1.1.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matlabengine==24.2.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
meshio==5.3.5
mistune==3.1.3
mkl==2025.0.1
mpmath==1.3.0
multimethod==1.12
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
plotly-express==0.4.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pymoo==0.6.1.3
pypardiso==0.4.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.1
PyWavelets==1.8.0
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-fem==11.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==78.1.0
simplejson==3.20.1
six==1.17.0
skorch==1.1.0
smmap==5.0.2
snakeviz==2.2.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.44.1
sympy==1.14.0
tabulate==0.9.0
tbb==2022.0.0
tcmlib==1.2.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
tqdm_joblib==0.0.4
traitlets==5.14.3
txt2tags==3.9
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.1
umf==0.9.1
uri-template==1.3.0
urllib3==2.3.0
visions==0.8.1
vtk==9.5.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xlrd==2.0.1
xlwings==0.33.14
ydata-profiling==4.16.1
